Modularized shrimp bot
Closes #24
