+++ shrimp_bot/views/index.pug
@@ -1,5 +1,4 @@
 extends layout
 
 block content
-  h1= title
-  p Welcome to #{title}
+  h1 Shrimpy!!
